# Allow usage of both args and opts to filter