Currently there is a problem when trying to use custom filter with both options and arguments. So executing following code:
```
defmodule MyCustomFilters do
  def asset_url(path, %{"version" => version, "case" => case} = _args, opts) do
    opts[:host] <> path <> "?v=" <> to_string(version) <> "&case=" <> case
  end
end

opts = [custom_filters: MyCustomFilters, host: "http://example.com"]

"{{ file_path | asset_url: version: 1, case: 'closed' }}"
|> Solid.parse!()
|> Solid.render!(%{ "file_path" => "/styles/app.css"}, opts)
|> IO.puts()
```

would fail.

This PR fixes that, refactoring a little bit custom filter calls and adding support for passing arguments and options as separate arguments to filters.

So given code works:
```
defmodule MyCustomFilters do
  def asset_url(path, %{"version" => version, "case" => case} = _args, opts) do
    opts[:host] <> path <> "?v=" <> to_string(version) <> "&case=" <> case
  end
end

opts = [custom_filters: MyCustomFilters, host: "http://example.com"]

"{{ file_path | asset_url: version: 1, case: 'closed' }}"
|> Solid.parse!()
|> Solid.render!(%{ "file_path" => "/styles/app.css"}, opts)
|> IO.puts()
# http://example.com/styles/app.css?v=1&case=closed
```

on sidenote it also adds support for ensuring that custom filter code is loaded.

Please note there isn't single change in existing tests, since it's all backward compatible.