@esse 2023-09-27
Note: after benchmarking I've removed the `Code.ensure_loaded/1` instruction from filters.ex

Here is my benchmark file:
```
defmodule XXX do
  def two(), do: 2
end

Benchee.run(
  %{
    "with ensure loaded" => fn -> Code.ensure_loaded(XXX); XXX.two() end,
    "without ensure_loaded" => fn -> XXX.two() end
  },
  time: 30,
  memory_time: 2
)
```

And results:
```
Operating System: macOS
CPU Information: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
Number of Available Cores: 12
Available memory: 32 GB
Elixir 1.13.2
Erlang 25.3.2.4

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 30 s
memory time: 2 s
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 1.13 min

Benchmarking with ensure loaded ...
Benchmarking without ensure_loaded ...

Name                            ips        average  deviation         median         99th %
without ensure_loaded        9.52 M      105.04 ns   ±494.44%         101 ns         127 ns
with ensure loaded           5.39 M      185.63 ns ±63960.27%         128 ns         185 ns

Comparison:
without ensure_loaded        9.52 M
with ensure loaded           5.39 M - 1.77x slower +80.59 ns

Memory usage statistics:

Name                     Memory usage
without ensure_loaded             0 B
with ensure loaded               24 B - ∞ x memory usage +24 B

**All measurements for memory usage were the same**
```