@edgurgel 2023-11-07
> Currently there is a problem when trying to use custom filter with both options and arguments. So executing following code:
> 
> ```
> defmodule MyCustomFilters do
>   def asset_url(path, %{"version" => version, "case" => case} = _args, opts) do
>     opts[:host] <> path <> "?v=" <> to_string(version) <> "&case=" <> case
>   end
> end
> 
> opts = [custom_filters: MyCustomFilters, host: "http://example.com"]
> 
> "{{ file_path | asset_url: version: 1, case: 'closed' }}"
> |> Solid.parse!()
> |> Solid.render!(%{ "file_path" => "/styles/app.css"}, opts)
> |> IO.puts()
> ```
> 
> would fail.

Could you elaborate on what would fail here? I ran the code above and got this as result:

```
http://example.com/styles/app.css?v=1&case=closed
```

What were you expecting?