@esse 2024-02-27
@edgurgel I checked and I can't reproduce this issue. Hence closing this PR - thanks for your input!